+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ New features
 * Add :code:`modeling.add_administration`
 * Add :code:`modeling.create_dataset_from_design`
 * Add :code:`modeling.read_datainfo`
+* Add :code:`modeling.write_datainfo`
 
 Changes
 =======